MF-1340 - Add CLI config TOML file (#1858)
* Add config


* Change key names


* Add config file path


* Add config file path


* Configure TOML parsing


* Add cli config command


* Remove debug log


* Use snake case


* Change from interactive command


* use map for keys


* Add cli logger level


* Use mainflux logger


* Fix cli


* Remove unnecessary comments


* Fix cli error handling


* Remove fmt


* Update config


* Fix cli


* Fix cli


* Fix cli


* Fix cli


* Modify CLI


* Fix errors


* Fix cli


* Fix errors


* Remove user token


* Remove unused variables


* Add empty line


* Add url parsing


* TEsts


* Make config path configurable


* Fix ci


* Remove empty toml


* Change url key identification


* Remove url parsing functions


* Handle parse error


* Handle url error


* Add marshal


* Update config


* Fix cli


* Handle file error


* Handle file error


* Modify url parsing


* Add usertoken


* Fix user token


* Fix errors


* Fix errors


* Remove string init


* Fix error


* Fix errors


* Remove config.toml from root


* Add empty line to config.toml


* Add empty line to config.toml


* Inline error handling


* Remove unnecessary type conversion


* Fix error handling


* Fix error handling


* Remove dynamic filters


---------

+++ cli/bootstrap.go
@@ -48,8 +48,8 @@ var cmdBootstrap = []cobra.Command{
 				return
 			}
 			pageMetadata := mfxsdk.PageMetadata{
-				Offset: uint64(Offset),
-				Limit:  uint64(Limit),
+				Offset: Offset,
+				Limit:  Limit,
 				State:  State,
 				Name:   Name,
 			}
